+++ ndk-build/src/apk.rs
@@ -190,8 +190,13 @@ impl Apk {
         }
     }
 
-    pub fn install(&self) -> Result<(), NdkError> {
+    pub fn install(&self, device_serial: Option<&str>) -> Result<(), NdkError> {
         let mut adb = self.ndk.platform_tool(bin!("adb"))?;
+
+        if let Some(device_serial) = device_serial {
+            adb.arg("-s").arg(device_serial);
+        }
+
         adb.arg("install").arg("-r").arg(&self.path);
         if !adb.status()?.success() {
             return Err(NdkError::CmdFailed(adb));
@@ -199,8 +204,13 @@ impl Apk {
         Ok(())
     }
 
-    pub fn start(&self) -> Result<(), NdkError> {
+    pub fn start(&self, device_serial: Option<&str>) -> Result<(), NdkError> {
         let mut adb = self.ndk.platform_tool(bin!("adb"))?;
+
+        if let Some(device_serial) = device_serial {
+            adb.arg("-s").arg(device_serial);
+        }
+
         adb.arg("shell")
             .arg("am")
             .arg("start")