+++ ndk-build/src/ndk.rs
@@ -377,9 +377,14 @@ impl Ndk {
         Err(NdkError::PlatformNotFound(min_sdk_version))
     }
 
-    pub fn detect_abi(&self) -> Result<Target, NdkError> {
-        let stdout = self
-            .platform_tool("adb")?
+    pub fn detect_abi(&self, device_serial: Option<&str>) -> Result<Target, NdkError> {
+        let mut adb = self.platform_tool("adb")?;
+
+        if let Some(device_serial) = device_serial {
+            adb.arg("-s").arg(device_serial);
+        }
+
+        let stdout = adb
             .arg("shell")
             .arg("getprop")
             .arg("ro.product.cpu.abi")